+++ src/main/java/bg/softuni/michevparquetsparquet/model/dto/ParquetDTO.java
@@ -9,7 +9,7 @@ public record ParquetDTO(
         Model model,
         Integer size,
         Integer classRate,
-        Integer price,
+        Double price,
         String imageUrl
 ) {
 }
